Выполнил задание - 2.16. Шаблонизируй это
НАКОНЕЦ))))).
