+++ src/mock/task.js
@@ -1,12 +1,4 @@
-
-import dayjs from 'dayjs';
-
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+import {getRandomInteger, takeDate} from '../utils.js';
 
 const descriptions = ['Lorem ipsum dolor sit amet, consectetur adipiscing elit.', 'Cras aliquet varius magna, non porta ligula feugiat eget.', 'Fusce tristique felis at fermentum pharetra.', 'Aliquam id orci ut lectus varius viverra.', 'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.', 'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.', 'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.', 'Sed sed nisi sed augue convallis suscipit in sed felis.', 'Aliquam erat volutpat.', 'Nunc fermentum tortor ac porta dapibus.', 'In rutrum ac purus sit amet tempus.'];
 
@@ -49,42 +41,8 @@ const generateCity = () => {
   return citys[randomIndex];
 };
 
-const generateDate = (date) => {
-  const maxDaysGap = 21520;
-  const maxDriveGap = 2880;
-  const isDate = (!date) ? dayjs().add(getRandomInteger(24, maxDaysGap), 'm') : dayjs(date).add(getRandomInteger(1, maxDriveGap), 'm');
-
-  return isDate;
-};
-
-const calculate = (d, h, m) => {
-  if (d !== 0) {
-    const day = dayjs().date(d).format('DD');
-    const hour = dayjs().hour(h).format('HH');
-    const minute = dayjs().minute(m).format('mm');
-
-    return `${day}D ${hour}H ${minute}M`;
-  } else if (h !== 0) {
-    const hour = dayjs().hour(h).format('HH');
-    const minute = dayjs().minute(m).format('mm');
-    return `${hour}H ${minute}M`;
-  } else {
-    const minute = dayjs().minute(m).format('mm');
-    return `${minute}M`;
-  }
-};
-
 export const generateTask = () => {
-  const dateStart = generateDate();
-  const eventDate = dayjs(dateStart).format('MMM DD');
-  const eventTimeStart = dayjs(dateStart).format('HH:mm');
-  const dateEnd = generateDate(dateStart);
-  const eventTimeEnd = dayjs(dateEnd).format('HH:mm');
-  const travelTimeMinute = dateEnd.diff(dateStart, 'm');
-  const travelTimeHour = dateEnd.diff(dateStart, 'h');
-  const travelTimeDay = dateEnd.diff(dateStart, 'd');
-
-  const travelTime = calculate(travelTimeDay, travelTimeHour, travelTimeMinute);
+  const {dateStart, eventDate, eventTimeStart, travelTime, eventTimeEnd} = takeDate();
 
   return {
     dateStart,
